feat: use address search from our backend instead of Digitransit API
Our backend has now an API endpoint for doing address searches in the
same way as the Digitransit address search API which was used before.
Our backend uses the same Digitransit API under the hood, and returns
the same kind of response but without many unnecessary fields.

This change was needed because the Digitransit API nowadays mandates API
key usage.